+++ src/game/entities/entities.h
@@ -6,7 +6,7 @@
 
 typedef struct entity_s entity;
 
-entity* create_entity(int x, int y, int width, int height);
+entity* create_entity(float x, float y, int width, int height);
 void free_entity(entity* e);
 
 SDL_Rect* get_pos(entity* e);
@@ -16,7 +16,7 @@ int get_damage(entity* e);
 
 void print_entity(entity* e);
 
-void set_pos(entity* e, int x, int y);
+void set_pos(entity* e, float x, float y);
 void set_sprite(entity* e, SDL_Texture* tex);
 void set_max_health(entity* e, int max_health);
 void set_health(entity* e, int health);
